requirements.txt
django==5.0
djangorestframework==3.14.0
python-dotenv==1.0.0
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.3
opencv-python==4.8.1.78
deepface==0.0.79
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
openai==1.3.5
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
django-cors-headers==4.6.0
django-allauth==65.3.0
dj-rest-auth==7.0.0
keras==3.7.0
